# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API definition for Go Web Boilerplate Server",
    "title": "Go Web Boilerplate Server",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "paths": {
    "/auth/credential": {
      "get": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Get user credential",
        "parameters": [
          {
            "type": "string",
            "description": "Authorization",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/edit": {
      "put": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Edit user",
        "parameters": [
          {
            "type": "string",
            "description": "Authorization",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "description": "EditUserRequest",
            "name": "EditUserRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.EditUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.EditUserResponse"
            }
          }
        }
      }
    },
    "/auth/forgotPassword": {
      "post": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Forgot password",
        "parameters": [
          {
            "description": "ForgotPasswordRequest",
            "name": "ForgotPasswordRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ForgotPasswordRequest"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Login user google",
        "parameters": [
          {
            "description": "LoginRequest",
            "name": "LoginRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.LoginResponse"
            }
          }
        }
      }
    },
    "/auth/loginGoogle": {
      "post": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Login user",
        "parameters": [
          {
            "description": "GoogleLoginRequest",
            "name": "GoogleLoginRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.GoogleLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.LoginResponse"
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Register new user",
        "parameters": [
          {
            "description": "CreateUserRequest",
            "name": "CreateUserRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.CreateUserResponse"
            }
          }
        }
      }
    },
    "/auth/resetPassword": {
      "post": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Reset password",
        "parameters": [
          {
            "description": "ResetPasswordRequest",
            "name": "ResetPasswordRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.ResetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ResetPasswordRequest"
            }
          }
        }
      }
    },
    "/healthcheck": {
      "get": {
        "description": "Put all mandatory parameter",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Healthcheck"],
        "summary": "Check system status",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.Status"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.CreateUserRequest": {
      "type": "object",
      "required": ["email", "fullname", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "fullname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.CreateUserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "fullname": {
          "type": "string"
        }
      }
    },
    "dto.EditUserRequest": {
      "type": "object",
      "properties": {
        "fullname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.EditUserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "fullname": {
          "type": "string"
        }
      }
    },
    "dto.ForgotPasswordRequest": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "dto.GoogleLoginRequest": {
      "type": "object",
      "required": ["token"],
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "dto.LoginRequest": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "dto.ResetPasswordRequest": {
      "type": "object",
      "required": ["password", "token"],
      "properties": {
        "password": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "dto.Status": {
      "type": "object",
      "properties": {
        "data": {},
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    }
  }
}